Consolidate outcome/state utils.
